# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Guideme Services
  description: HTTP service for managing the organizations and tutorials.
  version: ""
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /organization/multi_add:
    post:
      tags:
      - organization
      summary: multi_add organization
      description: Add n number of Organizations and return their IDs. This is a multipart
        request and each part has field name 'organization' and contains the encoded
        organization info to be added.
      operationId: organization#multi_add
      consumes:
      - multipart/form-data
      parameters:
      - name: array
        in: formData
        required: true
        schema:
          type: array
          items:
            $ref: '#/definitions/OrganizationRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              type: string
              example: Minima itaque vitae.
      schemes:
      - http
  /organization/multi_update:
    put:
      tags:
      - organization
      summary: multi_update organization
      description: Update Organizations with the given IDs. This is a multipart request
        and each part has field name 'organizations' and contains the encoded Organizations
        info to be updated. The IDs in the query parameter is mapped to each part
        in the request.
      operationId: organization#multi_update
      consumes:
      - multipart/form-data
      parameters:
      - name: ids
        in: query
        description: IDs of the Organizations to be updated
        required: true
        type: array
        items:
          type: string
        collectionFormat: multi
      - name: multi_update_request_body
        in: formData
        required: true
        schema:
          $ref: '#/definitions/OrganizationMultiUpdateRequestBody'
          required:
          - organizations
      responses:
        "204":
          description: No Content response.
      schemes:
      - http
...
definitions:
...
  OrganizationMultiUpdateRequestBody:
    title: OrganizationMultiUpdateRequestBody
    type: object
    properties:
      organizations:
        type: array
        items:
          $ref: '#/definitions/OrganizationRequestBody'
        description: Array of bottle info that matches the ids attribute
        example:
        - name: Blue's Cuvee
          url: http://www.google.com/
        - name: Blue's Cuvee
          url: http://www.google.com/
    example:
      organizations:
      - name: Blue's Cuvee
        url: http://www.google.com/
      - name: Blue's Cuvee
        url: http://www.google.com/
      - name: Blue's Cuvee
        url: http://www.google.com/
    required:
    - organizations
  OrganizationRequestBody:
    title: OrganizationRequestBody
    type: object
    properties:
      name:
        type: string
        description: Name of Organization
        example: Blue's Cuvee
        maxLength: 100
      url:
        type: string
        description: Company website URL
        example: http://www.google.com/
        pattern: (?i)^(https?|ftp)://[^\s/$.?#].[^\s]*$
    description: Organization describes an Organization to be stored.
    example:
      name: Blue's Cuvee
      url: http://www.google.com/
    required:
    - name
    - url
...